requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-cors==0.7.0
aiosignal==1.3.2
airportsdata==20241001
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.6.3
astor==0.8.1
astunparse==1.6.3
attrs==25.1.0
av==12.3.0
blake3==1.0.4
cachetools==4.2.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==3.1.1
colorful==0.5.6
compressed-tensors==0.9.0
contourpy==1.2.1
cycler==0.12.1
datasets==3.2.0
decorator==4.4.2
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
easyocr==1.7.1
einops==0.8.0
fastapi==0.115.7
filelock==3.17.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.6.1
fvcore==0.1.5.post20221221
gast==0.6.0
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.1
google-auth==2.38.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.66.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.24.6
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
interegular==0.3.3
iopath==0.1.10
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.5.0
kiwisolver==1.4.5
lark==1.2.2
lazy_loader==0.4
libclang==18.1.1
lm-format-enforcer==0.10.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mistral_common==1.5.2
ml-dtypes==0.3.2
mmcv-full==1.7.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-ml-py==12.570.86
openai==1.60.2
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
outlines==0.1.11
outlines_core==0.1.26
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
partial-json-parser==0.2.1.1.post5
pillow==10.4.0
platformdirs==4.2.2
portalocker==2.10.1
proglog==0.1.10
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post5
pycountry==24.6.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.1.2
pytesseract==0.3.13
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==13.7
pytorchvideo==0.1.5
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.41.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.7.1
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.6
six==1.16.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.1.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.8.10
tiktoken==0.7.0
timm==1.0.8
tokenizers==0.21.0
tomli==2.0.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.66.5
transformers==4.48.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ultralytics==8.2.81
ultralytics-thop==2.0.5
urllib3==2.2.2
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.1
-e git+https://github.com/vllm-project/vllm.git@f17f1d46086692a2973fad94860a95799fbd8582#egg=vllm
watchfiles==1.0.4
websockets==14.2
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.5.0
yacs==0.1.8
yapf==0.40.2
yarl==1.18.3
yt-dlp==2025.1.26
zipp==3.20.0
